% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clauseFn.R
\name{.tp}
\alias{.tp}
\alias{.ROM}
\alias{.aslope}
\alias{.slope}
\alias{.period}
\alias{.avgzscore}
\alias{.npeak}
\alias{.npit}
\title{Common \code{.f} functions used in clause}
\usage{
.tp(x)

.ROM(x)

.aslope(x)

.slope(x)

.period(x)

.avgzscore(x, mu, sigma)

.npeak(x)

.npit(x)
}
\arguments{
\item{x}{numeric vector of kinematics}
}
\value{
the computed statistics
}
\description{
Common \code{.f} functions used in clause
}
\seealso{
\link[=unistat]{clauses} for usage.
}
