% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{definitions}
\alias{definitions}
\alias{.dtIntoe}
\alias{.dtOutoe}
\alias{.dtEPS}
\alias{.dtDelayIncPkKnFx}
\alias{.dtDecPkKnFx}
\alias{.dtDelayDecPkKnFx}
\alias{.dtDecFxLoading}
\alias{.dtH2R}
\alias{.dtS2R}
\alias{.dtDecDF}
\alias{.dtDesc2R}
\alias{.dtIncPF}
\alias{.dtIncMaxDF}
\alias{.dtIncDF}
\alias{.dtIPP}
\alias{.dtN1R}
\alias{.dtFtD}
\alias{.dtDFSwg}
\alias{.dtAIR}
\alias{.dtIncFxIC}
\alias{.dtIncFxICEarlyKnExt}
\alias{.dtHypExt}
\alias{.dtFxMS}
\alias{.dtIncPkKnFx}
\alias{.dtDelayPkKnFx}
\alias{.dtDelayIncPkKnFx}
\alias{.dtDecPkKnFx}
\alias{.dtDelayDecPkKnFx}
\alias{.dtDecFxLoading}
\alias{.dtHipExtDef}
\alias{.dtIncHipFx}
\alias{.dtHipHypFx}
\alias{.dtIncHipExtMS}
\alias{.dtIncHipFxDecROM}
\alias{.dtDecHipFxIC}
\alias{.dtExHipAbdSwg}
\alias{.dtExHipAdd}
\alias{.dtExHipAbd}
\alias{.dtHipAddSt}
\alias{.dtHipIR}
\alias{.dtHipER}
\alias{.dtIncIRLateSt}
\alias{.dtIncPelTilt}
\alias{.dtDecPelTilt}
\alias{.dtDblBump}
\alias{.dtIncPelTiltIncROM}
\alias{.dtIncPelTiltIncROM}
\alias{.dtDecPelTiltIncROM}
\alias{.dtPelElevDepr}
\alias{.dtIncPelROM}
\alias{.dtProRetraction}
\alias{.dtIncPelRotROM}
\alias{.dtRevROM}
\title{Default feature definitions}
\usage{
definitions()

.dtIntoe(df)

.dtOutoe(df)

.dtEPS(df)

.dtDelayIncPkKnFx(df)

.dtDecPkKnFx(df)

.dtDelayDecPkKnFx(df)

.dtDecFxLoading(df)

.dtH2R(df)

.dtS2R(df)

.dtDecDF(df)

.dtDesc2R(df)

.dtIncPF(df)

.dtIncMaxDF(df)

.dtIncDF(df)

.dtIPP(df)

.dtN1R(df)

.dtFtD(df)

.dtDFSwg(df)

.dtAIR(df)

.dtIncFxIC(df)

.dtIncFxICEarlyKnExt(df)

.dtHypExt(df)

.dtFxMS(df)

.dtIncPkKnFx(df)

.dtDelayPkKnFx(df)

.dtDelayIncPkKnFx(df)

.dtDecPkKnFx(df)

.dtDelayDecPkKnFx(df)

.dtDecFxLoading(df)

.dtHipExtDef(df)

.dtIncHipFx(df)

.dtHipHypFx(df)

.dtIncHipExtMS(df)

.dtIncHipFxDecROM(df)

.dtDecHipFxIC(df)

.dtExHipAbdSwg(df)

.dtExHipAdd(df)

.dtExHipAbd(df)

.dtHipAddSt(df)

.dtHipIR(df)

.dtHipER(df)

.dtIncIRLateSt(df)

.dtIncPelTilt(df)

.dtDecPelTilt(df)

.dtDblBump(df)

.dtIncPelTiltIncROM(df)

.dtIncPelTiltIncROM(df)

.dtDecPelTiltIncROM(df)

.dtPelElevDepr(df)

.dtIncPelROM(df)

.dtProRetraction(df)

.dtIncPelRotROM(df)

.dtRevROM(df)
}
\arguments{
\item{df}{dataframe which has columns curve_id and 0,1,...,100, and (optionally) ofs (opposite foot strike time) for pelvis features.}
}
\value{
same as df but augmented with extra colunn(s) indicating whether the feature is detected
}
\description{
Default feature definitions
}
\details{
For help on writing your own feature definition see here.
}
\examples{
.dtIntoe(kinematics)
}
