% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detectAll}
\alias{detectAll}
\title{detect gait kinematics features}
\usage{
detectAll(x, plane = NULL, joint = NULL, comp = FALSE, fDict = NULL,
  .envir = parent.frame())
}
\arguments{
\item{x}{data.frame which consists (at least) of columns: joint, plane, curve_id, 0,1,2,...,100}

\item{plane}{a character vector. Only feature matching this plane will be detected.
Whatever is specified here should appear in the \code{fDict} data.frame.
For the default feature definitions, the possible choices are \code{"sag","cor","tra"}.
If \code{NULL} (default) all planes are detected.}

\item{joint}{a character vector. Only feature matching this joint will be detected. If \code{NULL} (default) all joints are detected.
#' Whatever is specified here should appear in the \code{fDict} data.frame.
For the default feature definitions, the possible choices are \code{"Pel","Hip","Knee","Ankle","FootProgress"}.
If \code{NULL} (default) all planes are detected.}

\item{comp}{logical. If \code{TRUE} the return value will have \code{attr(.,"comp")} which gives the clause result of the features.}

\item{fDict}{a data.frame (at least) with columns \code{fn},\code{joint}, and \code{plane}, used to decide which features to detect.
It should have a similiar format to that generated by \link{makeFDict}. If \code{NULL} (default) the default dictionary which comes with the package is used.}

\item{.envir}{an \code{environment} to supply. For normal interactive use, this argument can be ignored.
But if this function is to be used programmatically, it is possible that the custom made detectors cannot be found by this function.
In this case, simply supply the environment which contains the detectors using this argument.}
}
\value{
the same as x but augmented with extra columns indicating the detected features.
}
\description{
detect gait kinematics features
}
\details{
For individual feature definitions see \link[=.dtAIR]{definitions}.
}
\examples{
detectAll(kinematics)
detectAll(kinematics, plane="sag")

y=detectAll(kinematics, plane="sag", joint="Ankle",comp=T)
attr(y,"comp")
}
