% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRC.R
\name{getRC}
\alias{getRC}
\title{Get reference curve}
\usage{
getRC(.joint, .plane)
}
\arguments{
\item{.joint}{joint name}

\item{.plane}{plane name}
}
\value{
a dataframe of referene curve filtered to the specified joint and plane.
}
\description{
Get reference curve
}
\examples{
getRC("Ankle","sag")
}
\seealso{
rc
}
