% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PelR2L.R
\name{PelR2L}
\alias{PelR2L}
\title{Reconstruct Left Pelvis curves from Right curves}
\usage{
PelR2L(data, ofs, flipped = F)
}
\arguments{
\item{data}{dataframe with columns curve_id, side, 0,1,...,100, and whatever is specified by the \code{ofs} argument.}

\item{ofs}{the column name (quoted) of \code{data} which specifies the time point of opposite foot strike.}

\item{flipped}{logical. If TRUE the Right curves are flipped about the 0 degree axis.
This is needed when reconstructing the coronal and transverse plane pelvis curves, but not for the saggital curves.}
}
\value{
same as data but the right curves are reconstructed to look like the left curves.
}
\description{
Reconstruct Left Pelvis curves from Right curves
}
\details{
Since there is only 1 pelvis for each person, the information conveyed by the left and right curve should be the same.
However, the left and right curves do not typically look the same.
Therefore, it \emph{might} be useful to homogenize them so we are only looking at 1 set of patterns.

\code{PelR2L} reconstruct the left curves from the right curves.
It basically cuts out the second part of the curve, from the time of opposite (left) foot strike to the end of the gait cycle (t=100),
and paste it before the original beginning of the curve (t=0:t[opposite foot off]).
For the coronal and transverse plane, the whole curve can be flipped vertically about the 0 axis. i.e. multiplied by -1.
}
\examples{
## make some fake opposite foot strike time point
kinematics$ofs=sample(58:62,nrow(kinematics),replace=T)
homogenized=PelR2L(subset(kinematics,joint=="Pel"&plane=="sag"),"ofs")
matplot(t(homogenized\%>\%select(`0`:`100`)),type="l",lty=1,col=ifelse(temp$side=="R","blue","red"))
matplot(t(subset(kinematics,joint=="Pel"&plane=="sag")\%>\%select(`0`:`100`)),type="l",lty=1,col=ifelse(temp$side=="R","blue","red"))
PelR2L(subset(kinematics,joint=="Pel"&plane=="cor"),"ofs",flipped=TRUE)
}
