% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceCurves.R
\name{.rc}
\alias{.rc}
\alias{.tc}
\title{Reference Curves}
\usage{
.rc()

.tc()
}
\description{
\code{.rf()} returns a set of typically developed children's kinematics curves.
\code{.tc()} returns a set of typical feature curves (for a subset of features only).
}
