% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamp.R
\name{resamp101}
\alias{resamp101}
\title{Resample curves to t=0,...,100 using splines.}
\usage{
resamp101(x, ..., plot = FALSE)
}
\arguments{
\item{x}{data.frame with kinematics data in the wide format.}

\item{...}{column names (bare) of the kinematics data. It should be specified \link[dplyr]{select} style.}

\item{plot}{logical. If \code{TRUE} a plot of the observed values superimposed with the inteporalted values are generated. For large dataset the plot might be slow to generate and hard to interpret visually.}
}
\value{
same as \code{x} but \code{...} replaced by columns 0,1,...,100.
}
\description{
Resample curves to t=0,...,100 using splines.
}
\details{
The resampling (intepolation) is done using periodic cubic spline. In particular, \link{splinefun} is the workhorse function.
}
\examples{
X=kinematics\%>\%
filter(joint=="Knee",plane=="sag")\%>\%
  select(curve_id,`0`,`20`,`40`,`60`,`80`,`100`)
z=resamp101(X,`0`:`100`,plot=T)
}
